+++ src/gevent/tests/monkey_package/__init__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+"""
+Make a package.
+
+"""
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function