+++ src/gevent/tests/monkey_package/__main__.py
@@ -1,5 +1,5 @@
 from __future__ import print_function
-# This file makes this directory into a package.
+# This file makes this directory into a runnable package.
 # it exists to test 'python -m gevent.monkey monkey_package'
 print(__file__)
 print(__name__)