+++ src/gevent/threadpool.py
@@ -3,23 +3,25 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-import sys
 import os
+import sys
 
 from greenlet import greenlet as RawGreenlet
 
+from gevent import monkey
 from gevent._compat import integer_types
 from gevent.event import AsyncResult
 from gevent.exceptions import InvalidThreadUseError
 from gevent.greenlet import Greenlet
-from gevent.hub import _get_hub
+
+from gevent._hub_local import get_hub_if_exists
 from gevent.hub import _get_hub_noargs as get_hub
 from gevent.hub import getcurrent
 from gevent.hub import sleep
 from gevent.lock import Semaphore
 from gevent.pool import GroupMappingMixin
+from gevent.util import clear_stack_frames
 
-from gevent._threading import Lock
 from gevent._threading import Queue
 from gevent._threading import start_new_thread
 from gevent._threading import get_thread_ident
@@ -30,38 +32,138 @@ __all__ = [
     'ThreadResult',
 ]
 
+def _format_hub(hub):
+    if hub is None:
+        return '<missing>'
+    return '<%s at 0x%x thread_ident=0x%x>' % (
+        hub.__class__.__name__, id(hub), hub.thread_ident
+    )
 
 class _WorkerGreenlet(RawGreenlet):
     # Exists to produce a more useful repr for worker pool
     # threads/greenlets.
 
+    # Inform the gevent.util.GreenletTree that this should be
+    # considered the root (for printing purposes)
+    greenlet_tree_is_root = True
+
+    _thread_ident = 0
+    _exc_info = sys.exc_info
+    _get_hub_if_exists = staticmethod(get_hub_if_exists)
+    # We capture the hub each time through the loop in case its created
+    # so we can destroy it after a fork.
+    _hub_of_worker = None
+    # The hub of the threadpool we're working for. Just for info.
+    _hub = None
+
     def __init__(self, threadpool):
-        RawGreenlet.__init__(self, threadpool._worker)
-        self.thread_ident = get_thread_ident()
-        self._threadpool = threadpool
+        # Construct in the main thread (owner of the threadpool)
+        # The parent greenlet and thread identifier will be set once the
+        # new thread begins running.
+        RawGreenlet.__init__(self)
 
-        # Inform the gevent.util.GreenletTree that this should be
-        # considered the root (for printing purposes) and to
+        self._hub = threadpool.hub
+        # Avoid doing any imports in the background thread if it's not
+        # necessary (monkey.get_original imports if not patched).
+        # Background imports can hang Python 2 (gevent's thread resolver runs in the BG,
+        # and resolving may have to import the idna module, which needs an import lock, so
+        # resolving at module scope)
+        if monkey.is_module_patched('sys'):
+            stderr = monkey.get_original('sys', 'stderr')
+        else:
+            stderr = sys.stderr
+        self._stderr = stderr
+        # We can capture the task_queue; even though it can change if the threadpool
+        # is re-innitted, we won't be running in that case
+        self._task_queue = threadpool.task_queue
+
+        self._unregister_worker = threadpool._unregister_worker
+
+        threadpool._register_worker(self)
+        try:
+            start_new_thread(self._begin, ())
+        except:
+            self._unregister_worker(self)
+            raise
+
+    def _begin(self):
+        # we're in the new thread (but its root greenlet). Establish invariants and get going
+        # by making this the current greenlet.
+        self.parent = getcurrent() # pylint:disable=attribute-defined-outside-init
+        self._thread_ident = get_thread_ident()
         # ignore the parent attribute. (We can't set parent to None.)
-
-        self.greenlet_tree_is_root = True
         self.parent.greenlet_tree_is_ignored = True
+        try:
+            self.switch() # goto run()
+        except: # pylint:disable=bare-except
+            # run() will attempt to print any exceptions, but that might
+            # not work during shutdown. sys.excepthook and such may be gone,
+            # so things might not get printed at all except for a cryptic
+            # message. This is especially true on Python 2 (doesn't seem to be
+            # an issue on Python 3).
+            pass
 
-    @classmethod
-    def run_in_worker_thread(cls, threadpool):
-        # The root function of each new worker thread.
-        glet = cls(threadpool)
-        glet.switch()
+    def __fixup_hub_before_block(self):
+        hub = self._get_hub_if_exists() # Don't create one; only set if a worker function did it
+        if hub is not None:
+            hub.name = 'ThreadPool Worker Hub'
+            # While we block, don't let the monitoring thread, if any,
+            # report us as blocked. Indeed, so long as we never
+            # try to switch greenlets, don't report us as blocked---
+            # the threadpool is *meant* to run blocking tasks
+            if hub is not None and hub.periodic_monitoring_thread is not None:
+                hub.periodic_monitoring_thread.ignore_current_greenlet_blocking()
+            self._hub_of_worker = hub
 
-    def __repr__(self):
-        return "<ThreadPoolWorker at 0x%x thread_ident=0x%x %s>" % (
+    def __run_task(self, func, args, kwargs, thread_result):
+        try:
+            thread_result.set(func(*args, **kwargs))
+        except: # pylint:disable=bare-except
+            thread_result.handle_error((self, func), self._exc_info())
+        finally:
+            del func, args, kwargs, thread_result
+
+    def run(self):
+        # pylint:disable=too-many-branches
+        try:
+            while 1: # tiny bit faster than True on Py2
+                self.__fixup_hub_before_block()
+
+                task = self._task_queue.get()
+                try:
+                    if task is None:
+                        return
+
+                    self.__run_task(*task)
+                finally:
+                    task = None
+                    self._task_queue.task_done()
+        except Exception as e: # pylint:disable=broad-except
+            print("Failed to run worker thread", e, file=self._stderr)
+        finally:
+            # Re-check for the hub in case the task created it but then
+            # failed.
+            self.cleanup(self._get_hub_if_exists())
+
+    def cleanup(self, hub_of_worker):
+        if self._hub is not None:
+            self._hub = None
+            self._unregister_worker(self)
+            self._unregister_worker = lambda _: None
+            self._task_queue = None
+
+            if hub_of_worker is not None:
+                hub_of_worker.destroy(True)
+
+    def __repr__(self, _format_hub=_format_hub):
+        return "<ThreadPoolWorker at 0x%x thread_ident=0x%x hub=%s>" % (
             id(self),
-            self.thread_ident,
-            self._threadpool
+            self._thread_ident,
+            _format_hub(self._hub)
         )
 
+
 class ThreadPool(GroupMappingMixin):
-    # TODO: Document thread safety restrictions.
     """
     A pool of native worker threads.
 
@@ -120,11 +222,9 @@ class ThreadPool(GroupMappingMixin):
         # gevent.lock.Semaphore, this is only safe to use from a single
         # native thread.
         '_available_worker_threads_greenlet_sem',
-        # A native threading lock, used to protect internals
-        # that are safe to call across multiple threads.
-        '_native_thread_internal_lock',
-        # The number of running worker threads
-        '_num_worker_threads',
+        # A set of running or pending _WorkerGreenlet objects;
+        # we rely on the GIL for thread safety.
+        '_worker_greenlets',
         # The task queue is itself safe to use from multiple
         # native threads.
         'task_queue',
@@ -139,14 +239,19 @@ class ThreadPool(GroupMappingMixin):
         self.task_queue = Queue()
         self.fork_watcher = None
 
+        self._worker_greenlets = set()
         self._maxsize = 0
         # Note that by starting with 1, we actually allow
         # maxsize + 1 tasks in the queue.
         self._available_worker_threads_greenlet_sem = Semaphore(1, hub)
-        self._native_thread_internal_lock = Lock()
-        self._num_worker_threads = 0
         self._set_maxsize(maxsize)
         self.fork_watcher = hub.loop.fork(ref=False)
+
+    def _register_worker(self, worker):
+        self._worker_greenlets.add(worker)
+
+    def _unregister_worker(self, worker):
+        self._worker_greenlets.discard(worker)
 
     def _set_maxsize(self, maxsize):
         if not isinstance(maxsize, integer_types):
@@ -172,12 +277,13 @@ class ThreadPool(GroupMappingMixin):
     will block waiting for a task to finish.
     """)
 
-    def __repr__(self):
-        return '<%s at 0x%x %s/%s/%s hub=<%s at 0x%x thread_ident=0x%s>>' % (
+    def __repr__(self, _format_hub=_format_hub):
+        return '<%s at 0x%x tasks=%s size=%s maxsize=%s hub=%s>' % (
             self.__class__.__name__,
             id(self),
             len(self), self.size, self.maxsize,
-            self.hub.__class__.__name__, id(self.hub), self.hub.thread_ident)
+            _format_hub(self.hub),
+        )
 
     def __len__(self):
         # XXX just do unfinished_tasks property
@@ -186,9 +292,7 @@ class ThreadPool(GroupMappingMixin):
         return self.task_queue.unfinished_tasks
 
     def _get_size(self):
-        # TODO: This probably needs to acquire the lock. We
-        # modify this from self._add_thread under a lock.
-        return self._num_worker_threads
+        return len(self._worker_greenlets)
 
     def _set_size(self, size):
         if size < 0:
@@ -197,17 +301,17 @@ class ThreadPool(GroupMappingMixin):
             raise ValueError('Size of the pool cannot be bigger than maxsize: %r > %r' % (size, self._maxsize))
         if self.manager:
             self.manager.kill()
-        while self._num_worker_threads < size:
+        while len(self._worker_greenlets) < size:
             self._add_thread()
         delay = self.hub.loop.approx_timer_resolution
-        while self._num_worker_threads > size:
-            while self._num_worker_threads - size > self.task_queue.unfinished_tasks:
+        while len(self._worker_greenlets) > size:
+            while len(self._worker_greenlets) - size > self.task_queue.unfinished_tasks:
                 self.task_queue.put(None)
             if getcurrent() is self.hub:
                 break
             sleep(delay)
             delay = min(delay * 2, .05)
-        if self._num_worker_threads:
+        if self._worker_greenlets:
             self.fork_watcher.start(self._on_fork)
         else:
             self.fork_watcher.stop()
@@ -226,13 +330,32 @@ class ThreadPool(GroupMappingMixin):
 
     def _on_fork(self):
         # fork() only leaves one thread; also screws up locks;
-        # let's re-create locks and threads.
+        # let's re-create locks and threads, and do our best to
+        # clean up any worker threads left behind.
         # NOTE: See comment in gevent.hub.reinit.
         pid = os.getpid()
         if pid != self.pid:
-            # The OS threads have been destroyed, but the Python objects
-            # may live on, creating refcount "leaks"
+            # The OS threads have been destroyed, but the Python
+            # objects may live on, creating refcount "leaks". Python 2
+            # leaves dead frames (those that are for dead OS threads)
+            # around; Python 3.8 does not.
+            thread_ident_to_frame = dict(sys._current_frames())
+            for worker in list(self._worker_greenlets):
+                frame = thread_ident_to_frame.get(worker._thread_ident)
+                clear_stack_frames(frame)
+                worker.cleanup(worker._hub_of_worker)
+                # We can't throw anything to the greenlet, nor can we
+                # switch to it or set a parent. Those would all be cross-thread
+                # operations, which aren't allowed.
+                worker.__dict__.clear()
+
+            # We've cleared f_locals and on Python 3.4, possibly the actual
+            # array locals of the stack frame, but the task queue may still be
+            # referenced if we didn't actually get all the locals. Shut it down
+            # and clear it before we throw away our reference.
+            self.task_queue.kill()
             self.__init__(self._maxsize)
+
 
     def join(self):
         """Waits until all outstanding tasks have been completed."""
@@ -247,14 +370,14 @@ class ThreadPool(GroupMappingMixin):
 
     def _adjust_step(self):
         # if there is a possibility & necessity for adding a thread, do it
-        while (self._num_worker_threads < self._maxsize
-               and self.task_queue.unfinished_tasks > self._num_worker_threads):
+        while (len(self._worker_greenlets) < self._maxsize
+               and self.task_queue.unfinished_tasks > len(self._worker_greenlets)):
             self._add_thread()
         # while the number of threads is more than maxsize, kill one
         # we do not check what's already in task_queue - it could be all Nones
-        while self._num_worker_threads - self._maxsize > self.task_queue.unfinished_tasks:
+        while len(self._worker_greenlets) - self._maxsize > self.task_queue.unfinished_tasks:
             self.task_queue.put(None)
-        if self._num_worker_threads:
+        if self._worker_greenlets:
             self.fork_watcher.start(self._on_fork)
         elif self.fork_watcher is not None:
             self.fork_watcher.stop()
@@ -263,27 +386,20 @@ class ThreadPool(GroupMappingMixin):
         delay = 0.0001
         while True:
             self._adjust_step()
-            if self._num_worker_threads <= self._maxsize:
+            if len(self._worker_greenlets) <= self._maxsize:
                 return
             sleep(delay)
             delay = min(delay * 2, .05)
 
     def adjust(self):
         self._adjust_step()
-        if not self.manager and self._num_worker_threads > self._maxsize:
+        if not self.manager and len(self._worker_greenlets) > self._maxsize:
             # might need to feed more Nones into the pool to shutdown
             # threads.
             self.manager = Greenlet.spawn(self._adjust_wait)
 
     def _add_thread(self):
-        with self._native_thread_internal_lock:
-            self._num_worker_threads += 1
-        try:
-            start_new_thread(_WorkerGreenlet.run_in_worker_thread, (self,))
-        except:
-            with self._native_thread_internal_lock:
-                self._num_worker_threads -= 1
-            raise
+        _WorkerGreenlet(self)
 
     def spawn(self, func, *args, **kwargs):
         """
@@ -337,69 +453,6 @@ class ThreadPool(GroupMappingMixin):
             semaphore.release()
             raise
         return result
-
-    def _decrease_size(self):
-        if sys is None:
-            return
-        _lock = self._native_thread_internal_lock
-        if _lock is not None: # XXX: When would this be None?
-            with _lock:
-                self._num_worker_threads -= 1
-
-    def __ignore_current_greenlet_blocking(self, hub):
-        if hub is not None and hub.periodic_monitoring_thread is not None:
-            hub.periodic_monitoring_thread.ignore_current_greenlet_blocking()
-
-    def _worker(self):
-        # pylint:disable=too-many-branches
-        need_decrease = True
-        try:
-            while 1: # tiny bit faster than True on Py2
-                h = _get_hub() # Don't create one; only set if a worker function did it
-                if h is not None:
-                    h.name = 'ThreadPool Worker Hub'
-                task_queue = self.task_queue
-                # While we block, don't let the monitoring thread, if any,
-                # report us as blocked. Indeed, so long as we never
-                # try to switch greenlets, don't report us as blocked---
-                # the threadpool is *meant* to run blocking tasks
-                self.__ignore_current_greenlet_blocking(h)
-                task = task_queue.get()
-                try:
-                    if task is None:
-                        need_decrease = False
-                        self._decrease_size()
-                        # we want first to decrease size, then decrease unfinished_tasks
-                        # otherwise, _adjust might think there's one more idle thread that
-                        # needs to be killed
-                        return
-                    func, args, kwargs, thread_result = task
-                    try:
-                        value = func(*args, **kwargs)
-                    except: # pylint:disable=bare-except
-                        exc_info = getattr(sys, 'exc_info', None)
-                        if exc_info is None:
-                            return
-                        thread_result.handle_error((self, func), exc_info())
-                    else:
-                        if sys is None:
-                            return
-                        thread_result.set(value)
-                        del value
-                    finally:
-                        del func, args, kwargs, thread_result, task
-                finally:
-                    if sys is None:
-                        return # pylint:disable=lost-exception
-                    task_queue.task_done()
-        finally:
-            if need_decrease:
-                self._decrease_size()
-            if sys is not None:
-                hub = _get_hub()
-                if hub is not None:
-                    hub.destroy(True)
-                del hub
 
     def _apply_immediately(self):
         # If we're being called from a different thread than the one that
@@ -527,22 +580,21 @@ else:
     from gevent._util import Lazy
     from concurrent.futures import _base as cfb
 
-    def _wrap_error(future, fn):
+    def _ignore_error(future_proxy, fn):
         def cbwrap(_):
             del _
-            # we're called with the async result, but
-            # be sure to pass in ourself. Also automatically
-            # unlink ourself so that we don't get called multiple
-            # times.
+            # We're called with the async result (from the threadpool), but
+            # be sure to pass in the user-visible _FutureProxy object..
             try:
-                fn(future)
+                fn(future_proxy)
             except Exception: # pylint: disable=broad-except
-                future.hub.print_exception((fn, future), *sys.exc_info())
+                # Just print, don't raise to the hub's parent.
+                future_proxy.hub.print_exception((fn, future_proxy), None, None, None)
         return cbwrap
 
-    def _wrap(future, fn):
+    def _wrap(future_proxy, fn):
         def f(_):
-            fn(future)
+            fn(future_proxy)
         return f
 
     class _FutureProxy(object):
@@ -553,7 +605,6 @@ else:
 
         @Lazy
         def _condition(self):
-            from gevent import monkey
             if monkey.is_module_patched('threading') or self.done():
                 import threading
                 return threading.Condition()
@@ -604,10 +655,11 @@ else:
                 raise concurrent.futures.TimeoutError()
 
         def add_done_callback(self, fn):
+            """Exceptions raised by *fn* are ignored."""
             if self.done():
                 fn(self)
             else:
-                self.asyncresult.rawlink(_wrap_error(self, fn))
+                self.asyncresult.rawlink(_ignore_error(self, fn))
 
         def rawlink(self, fn):
             self.asyncresult.rawlink(_wrap(self, fn))